moved solution to solution folder
