+++ go/src/solution/1-hello/hello.go
