+++ go/src/solution/1-hello/hello_test.go
